TodoItem: Fix failing UI test by removing extra whitespace
Removes a leading whitespace character and replaces it with CSS padding.

After the previous commit, which updates a UI test to actually do its job, the
UI test actually failed. This is because there was an `nbsp;` which was meant
to be removed once TailwindCSS was installed, but it was forgotten. This
caused new tasks to have a leading whitespace and thus the UI test would fail.
Turns out UI testing can be useful?